Add package logwatch-svn
The purpose of creating logwatch-svn, instead of using logwatch in
community repo, is that logwatch has not been frequently released by the
upstream. The latest release 7.4.0 is made public in March 2011, and there
are over 100 commits(from r47 to r157) in the svn trunk ever since then.
Users would benefit from these new features in the svn trunk with no doubt.

Signed-off-by: Lance Chen <cyen0312@gmail.com>